requirements.txt
openai-whisper
yt-dlp
langchain
langchain-openai
langchain-community
faiss-cpu
streamlit
python-dotenv
